Add UIKitKeyEventConversions
