+++ Sources/OpenSwiftUI/Event/Event/ModifierKeysChanged.swift
@@ -6,6 +6,10 @@
 //  Status: WIP
 //  ID: FC441B6AC7E06AA4A028C48BD48D4AD4 (SwiftUI)
 
-//extension EventModifiers {
-//    fileprivate static let _modifierKeySource: any ModifierKeySource
-//}
+// TODO: EventModifiers.AsyncValues
+
+// TODO: OnModifierKeysChangedModifier
+
+extension EventModifiers {
+    fileprivate static let _modifierKeySource: any ModifierKeySource = DefaultModifierKeySource()
+}
