+++ Sources/OpenSwiftUI/Event/InputEvent/KeyboardShortcut/KeyboardShortcutBinding.swift
@@ -1,0 +1,9 @@
+//
+//  KeyboardShortcutBinding.swift
+//  OpenSwiftUI
+
+struct KeyboardShortcutBinding {
+    var shortcut: KeyboardShortcut
+    var action: ()->Void
+    var title: String?
+}
