+++ Sources/OpenSwiftUI/Event/Platform/UIKit/UIKitKeyPressResponder.swift
@@ -113,58 +113,4 @@ final class UIKitKeyPressResponder: UIResponder {
 protocol FallbackResponderProvider: AnyObject {
     var defaultNextResponder: UIResponder? { get }
 }
-
-// FIXME
-
-extension UIPress: KeyEventProvider {
-    func resolve(phase: EventPhase) -> KeyEvent? {
-        guard let key else {
-            return nil
-        }
-        return KeyEvent(
-            phase: phase,
-            timestamp: Time(seconds: timestamp),
-            modifiers: EventModifiers(key.modifierFlags),
-            keys: keyEquivalentString(key.charactersIgnoringModifiers),
-            stringValue: keyEquivalentString(key.characters),
-            keyID: AnyHashable(key.keyCode)
-        )
-    }
-}
-
-// MARK: - Key Input Conversion
-
-private let keyInputToKeyEquivalentMap: [String: Character] = [
-    UIKeyCommand.inputUpArrow: "\u{F700}",
-    UIKeyCommand.inputDownArrow: "\u{F701}",
-    UIKeyCommand.inputLeftArrow: "\u{F702}",
-    UIKeyCommand.inputRightArrow: "\u{F703}",
-    UIKeyCommand.inputEscape: "\u{001B}",
-    UIKeyCommand.inputDelete: "\u{0008}",
-    UIKeyCommand.inputPageUp: "\u{F72C}",
-    UIKeyCommand.inputPageDown: "\u{F72D}",
-    UIKeyCommand.inputHome: "\u{F729}",
-    UIKeyCommand.inputEnd: "\u{F72B}",
-    UIKeyCommand.f1: "\u{F704}",
-    UIKeyCommand.f2: "\u{F705}",
-    UIKeyCommand.f3: "\u{F706}",
-    UIKeyCommand.f4: "\u{F707}",
-    UIKeyCommand.f5: "\u{F708}",
-    UIKeyCommand.f6: "\u{F709}",
-    UIKeyCommand.f7: "\u{F70A}",
-    UIKeyCommand.f8: "\u{F70B}",
-    UIKeyCommand.f9: "\u{F70C}",
-    UIKeyCommand.f10: "\u{F70D}",
-    UIKeyCommand.f11: "\u{F70E}",
-    UIKeyCommand.f12: "\u{F70F}",
-]
-
-private func keyEquivalentString(_ input: String) -> String {
-    guard let keyEquivalent = keyInputToKeyEquivalentMap[input] else {
-        return input
-    }
-    return String(keyEquivalent)
-}
-
-
 #endif
